API/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.1
